+++ webpack.config.js
@@ -5,12 +5,36 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 const __dirname = fileURLToPath(import.meta.url);
 
-let config = {
+import resolveTypeScriptPluginModule from 'resolve-typescript-plugin';
+const ResolveTypeScriptPlugin = resolveTypeScriptPluginModule.default;
+
+const srcPath = path.resolve(__dirname, '../src');
+const entryPath = './docs/assets/js/src';
+const distPath = path.resolve(__dirname, '../docs/assets/js/dist');
+
+const config = {
 	mode: process.env.MODE,
-	entry: './docs/assets/js/src/main.js',
+	entry: {
+		'docs-script': `${entryPath}/docs-script.ts`,
+	},
 	output: {
-		path: path.resolve(__dirname, '../docs/assets/js/dist'),
-		filename: 'bundle.js',
+		path: distPath,
+		filename: '[name].bundle.js',
+	},
+	resolve: {
+		fullySpecified: true,
+		plugins: [new ResolveTypeScriptPlugin()],
+		alias: {
+			'@cipscis/debounce': `${srcPath}/debounce.ts`,
+		},
+	},
+	module: {
+		rules: [
+			{
+				test: /\.ts$/,
+				loader: 'ts-loader',
+			},
+		],
 	},
 };